requirements.txt
django==1.8.1
django-braces
django-widget-tweaks
django-mathfilters
djangorestframework
django-debug-toolbar
django-celery
requests
psutil
wget
pytz
six
sqlparse
numpy
scipy
pytest-django
nose
